+++ include/ALGraphV2.h
@@ -11,12 +11,12 @@
 
 // Adjacency list implementation of Graph
 class ALGraphV2{
-    int v, e;
+    uint64_t v, e;
     std::vector<int>* edges;
     std::vector<int>* weights;
 
 public:
-    ALGraphV2(int v, int e) : v(v), e(e){
+    ALGraphV2(uint64_t v, uint64_t e) : v(v), e(e){
         edges = new std::vector<int>[v + 2];
         weights = new std::vector<int>[v + 2];
         for(int i = 0; i <= v; ++i)
